added a 'add new alarm' screen and the backend behind said screen
the user can now go to a groups page and click a button to add a new alarm. This button opens a new screen where a name, date and time can be selected. When everything is filled in, a 'confirm' button can be pressed, which sends the data to the database, and puts the user back into the screen of the group they were in