added a basic alarm function
This alarm function gets called by the background workmanager class that runs every 15 minutes. If an alarm exists in the database, that is within 20 minutes (for possible lateness of the worker class) the alarm function gets made. This function runs at the set time and can also run from the background. Clicking the dismiss button opens up the app (more functionality can be added in the future)